me voa dormir

+++ tumba/models.py
@@ -16,8 +16,8 @@ class Lote (BaseModelTumba):
     x = models.FloatField(default=0.1,verbose_name='x', blank="true")  # Coordenada X
     y = models.FloatField(default=0.1,verbose_name='y', blank="true")  # Coordenada Y
     rotation = models.FloatField(default=0.1, verbose_name='rotacion', blank="true")  # Rotación opcional
-    text_x = models.FloatField(default=0.0,verbose_name='text x', blank="true")  # Posición X del texto
-    text_y = models.FloatField(default=0.0,verbose_name='text y', blank="true")  # Posición Y del texto
+    text_x = models.IntegerField(default=0,verbose_name='text x', blank="true")  # Posición X del texto
+    text_y = models.IntegerField(default=0,verbose_name='text y', blank="true")  # Posición Y del texto
     trans_r_x = models.FloatField(default=0.1, verbose_name='Translate X', blank=True)  # Coordenada X del translate rectangulo
     trans_r_y = models.FloatField(default=0.1, verbose_name='Translate Y', blank=True)  # Coordenada Y del translate rectangulo
     trans_t_x = models.FloatField(default=0.1, verbose_name='Translate X', blank=True)  # Coordenada X del translate texto
